Fix missing item map issue (#68)
Always show both the map and the table if we have multiple items.
Owners we don't know about don't appear on the map. If we have
only one marker, don't show the table, as the info is visible in
the map by default (if we have one marker only, it is zoomed in).