Support paging for details view based on last result list (#130)
Store hbz IDs and URL of the last result list in the session.
Store previous and next hbz ID on access of details view in the
session. Use all three to create previous and next pager links and
link to last result list at the bottom of the details view.

Large data (previous and next ID for each ID, the list of IDs in
the last search result) is not stored in the session directly
(too large for the session), but in the cache, using a unique
session ID. Small data (current `previous` and `next` ID, URL to
the last search) are stored directly in the session.

To get correct pagination in the details view, individual detail
pages are removed from the cache when they are part of a new search
result, but stay cached when paging back and forth.