Rework tests
- merge test classes into one class
- rename test class
- test ETL and outcome of JSON-LD and NTriples
- add external json-ld context URI
- remove single ntriple files and use xmlClob.tar.bz2 instead
- remove singe hbz01.json test file
- add 369 jsonld test files as they appear in ES as records
- change online test to use server gaia

Executing non-online tests the JSON-LD context is stored within every
JSON-LD document in elasticsearch. The context resides in the package
"lobid-rdf-to-json" These are compared to JSON-LD files.
For the ease of reading the documents these are ripped of the context
and given an external context URI. The online test however makes indeed
use of the external context, given to the EtikettMaker to compute the
JSON-LD.