Check if test file is XML
If a resource that is in the list "testIds.txt" is removed from the catalog
it can no more be get via http://lobid.org/resource?id=$id&format=source.
The result of such a lookup is "No source data found for request" which mustn't
be saved as an XML file because that parsing will result in:
 org.xml.sax.SAXParseException; lineNumber: 1; columnNumber: 1; Content is not allowed in prolog
This sanity check makes sure that empty resources are not stored.

- the repo's directory is no more left to avoid confusion

See https://github.com/hbz/lobid-resources/pull/23#issuecomment-184697253.