Parse doi, ignoring inproper prefix and suffix
If a doi is not valid the consequential building of an http-doi would result
in a broken URI, thus the whole document would be dropped when indexing it.
Better is to just ignore the wrongly catalogued doi and index the rest of the
document.
This is done with a simple regex.

- add test resource HT018939763

See #85.