Admit objects to be URIs even if they are not
Allowing invalid URIs to be URIs will allow to follow the expected
field structure in the ES index profile (URIs come with subfield "id").
This avoids the feared "org.elasticsearch.index.mapper.MapperParsingException",
e.g. "object mapping [fulltextOnline] trying to serialize a value with no field
associated with it". which would result in the dumping of the whole document.

See #85.

- add resource HT016618741