Support drill down within a facet (#336)
Include current facet selection when getting other facet values.
Makes the facet self restricting: it shows only values that appear
with the current selection. Makes the facet behavior more
consistent, as facets in general narrow down the current results.

Also fixes the problem of missing selected values which were not
part of the top results with the previous behavior.