Write test files automatically if the data changes
If the computated data differs from the expected data it becomes
the new expected data. A "git diff" then shows the effects on the
data. The files can then be easily commited.

- reduce complexity in test class

Instead of using base name and some addendum and the relative file
just use the file. This comes also very handy when logging messages
because the output of the filenames now also comes with their pathes
so that one can see which file actually is used.

- fail if to be tested files are not of equal size

Before this commit an expected result greater than
the computed one would not be recognised as a dissimilarity.

- write newly generated files as new test files
- set "failing only once at the end" as default

Both changes make it easier to make a complete diff and a complete
commit.

- update test set

The three commits including and before of c3dfac31d6d69ddf814d0812c4bb5587b991f57a
lacked the update of the tests.

See #30.