Add a label to every object
Guarantees a label as sidecar for every URI.
First fallback: If an etikett is configured but "label" is missing the "name"
will be taken. Second fallback: the domainname (and a possible path) will be
extracted from the URI. This domainname is lookuped in the labels.
Last fallback: If there is nothing found, the domainname itself will be the label.

See hbz/lobid-resources#132.