Adjust tests: first overwrite all files, then fail
If one of the tests hits an unexepcted result, the test failed immediatley.
In contrast to this the other tests work like this:
first test and update all test files (if some were failing), than return the
"fail" status at the end. So if you would run the tests again it would end
with "success" and the changed test files are comparable with "git diff".

This commit adjusts the behaviour of the test to those of the other tests.

- log some more details when comparing json
- update test files resulting from the commit before this one