Update context location
Note that the productive context for the test files is used.

If a test set shall be indexed this is done by running "buildAndETLTestSet.sh"
which will use the context pointing to test.lobid.org. You should run this in
the proper repo so that the context properly resolves.

See #237.