Add label of item
This information is redundant because the field name "callNumber" already
has the exact information. But then, the "label" yields a different semantic.
Also, as every id has a label, the removing of the label at a whole would break
the consistency of the data structure.

See https://github.com/hbz/lobid-resources/pull/416#issuecomment-298641289 point 1.