Avoid unnecessary fallback-lookups
When a fallback geodata-lookup was done once for a query the lookup
should not be done for the same query again being in the same transformtion
session, because when the query fails at the first time it will most probably
fail subsequently.

This does not only prevent repeating lookups of failed queries but also of
successfully queries which were stored in the geodata-elasticsearch-index
but which scores not enough so that the fallback lookup would be done
repeatedly.

See #579.