Add a workaround to avoid a "double" separator bug
This is necessary for #617 to build on travis.

Some RDF library converts a json double (using a '.' as decimal separator) to
a number with a ',' separator.
Some transitive library is doing this. As travis obviously takes precedence of
an other class as our local builds (which were all good) this was weired.

But as this is conversion is just used internally for testing purpose, a simple
workaround (replacing the wrong separator manually ) is sufficient.

See #616.