Restrict dct:title to be of type "literal"
By default the decision if an rdf-object is a literal or another node is
determined by the value of this rdf-object: if it begins with a URI-scheme
its treated as another node, else as literal.
Resources having a title beginning with "http" were droppped because that
would make a node while elasticsearch cannot allow different field-structures
for the same field-name (one as a plain literal, the other as a json-object).

See #625.