Fix set refresh intervall to stop while indexing
The former setting has only set the refresh intervall to -1 for every bulk
indexing process (of which ca. 12k (5k docs per bulk) were done when indexing
80M docs).
The new setting is true for the whole indexing process reducing the refreshs
to ideally 1 at the end of the process (well, there will be more when RAM is
exceeded and the data is flushed to disk).
Also, replicas is set to 0 so that only once at the end of the index process
these are generated, minimizing disk oprations, thus reducing the load thus
increasing free hardware resources which are then free to query other indices
in parallel.

- don't update the geo-cache when indexing test-resources
- set cluster name to ES5.6-cluster

See #648.