Fix context issue in context-labels.json
The actual context is generated from context-labels.json, so this
is where this has to be fixed. Also, the correct fix is not to make
the existing `location` key uppercase (which is used elsewhere as a
property), but to add a new `Location` key to be used as `type`.

See https://github.com/hbz/lobid-resources/issues/652