Add self-reference always as simple json-object
Cyclic graphs are no good. So references were put as simple values to the json.
This results in a contradiction when badly catalogued data is cyclic (see also
 #665) since elasticsearch index needs to be agreed on either defining a field
as value or as object, not both. As we degreed to provide always an "id" and a
"label" when it comes to URIs this commit enforces our agreement .
Mind that "itemOf" wasn't an object before but is now.

See #503.

- add TT000136775