Create new index "deleted"
This provides a new index to lookup deleted hbz01-resources. It works like this:
So called "Loeschsaetze" in aleph Mab xml are stored in an own index called
"deleted". As each Loeschsatz only contains an internal aleph system number
a lookup to associate this internal id with a lobid-resources id is needed.
Since lobid/lodmill#258 the API at http://lobid.org/hbz01 provides a lookup
service for aleph internal sysnumbers. This API is used to enrich the minimal
data residing in the Loeschsatz. As even no date is provided to specify when
the resource was deleted this date is set in the morph. The date used is just
the date at when the transformation is done, so this is not very accurate but
the best available. Note: it may sometimes be necessary to reindex a Loeschsatz
at a later point, say some days. The date would then become inaccurate, but
again - it's all we have.

- parametrize the classes to control which and how an index should be build
- improve logging by catching and printing more exceptions
- enrich loeschsatz with data gained from hbz01-API
- provide index-config and a deletion-labels directory for new jsonld-context
- provide junit test for a Loeschsatz

See #258.