Remove stopping process in "start"
Stopping a process and starting it is rather a "restart"
than a "start". For "restart" we have a script on it's own
called "restart.sh".

So calling "restart.sh" caused two times a killing of the process
(in fact 4 times since it was as fallback a "kill -9" implemented).

As we experienced again problems with monit restarting the process
this may be the cause of it.