Add more convenient multithreading
The ObjectThreader divides incoming objects and passes them to receivers which
are running in their own threads. The receivers can be added like in the Tee
receivers.

This is more convenient then to use the ObjectDivider chained to the
ObjectPipeDecoupler.

Depends on some methods of org/culturegraph/mf/framework/DefaultTee.java to
be not final.

See #967.