Add sanity check before updating the index aliases
The "deletedCountTest" is facultative because it depends on the synchrony
of the creation of the fulldump and the creation of the index (must be the same
day). So if e.g. in the middle of the week a new staging index is build with an
4 day old fulldump the test may probably fail. So use "deletedCountTest" only
when doing the main etl procedure at the weekend.

See #975.