Add Webhook "switchalias"
Switches aliases as defined in webhook.alma.basedump.indexname and that
same index name suffixing a "-staging", if the latter is newer, thus switch
a "-staging" to be productive by removeing that "-staging"-suffix and make
the former productive alias to staging by suffixing a "-staging" to it.
Makes some sanity checks (amount of documents and size of index).

Invoked with "token" paramter. If that token is not ok, switching is declined.
Sends an email to recepients configured in resources.conf.

- configure routes

See #1213.