Fix tests (#1360)
Methods of ObjectMapper removes the '\n' at the end of strings so that
the compared string of the test files are not equal (for they have (and
keep) their appended '\n').
By appending the '\n' at the end of the generated json it compares to equal
with the test files.

- reformat code
- remove the hard set of "generateTestData" in test code