Ignore temporary callnumber in permanent location.
In #1803 (eab045c), the temporary callnumber was introduced as the preferred callnumber (as per the decision by FEx Daten). But some local workflows keep the temporary callnumber even when the item is removed from the temporary location and transferred back to the permanent location, which has been reported in hbz-Support-Ticket#2346067. Hence, the temporary callnumber must not be considered in those cases.

In Alma, the transfer is recorded through a simple switch ("Item is in temporary location" = Yes/No). But this information is not transported in the Alma Publishing.